+++ routes/web.php
@@ -10,10 +10,6 @@ Route::get('/', function () {
     return redirect()->route('tenants.index');
 });
 
-/* Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard'); */
-
 Route::middleware('auth')->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
@@ -29,6 +25,9 @@ Route::middleware('auth')->group(function () {
     Route::get('/module-requests', [ModuleRequestController::class, 'index'])->name('module-requests.index');
     Route::post('/module-requests/{moduleRequest}/approve', [ModuleRequestController::class, 'approve'])->name('module-requests.approve');
     Route::post('/module-requests/{moduleRequest}/reject', [ModuleRequestController::class, 'reject'])->name('module-requests.reject');
+
+    // Domain
+    
 });
 
 require __DIR__.'/auth.php';